+++ package.json
@@ -16,7 +16,7 @@
     "express": "^4.18.1",
     "framer-motion": "^6.3.10",
     "github-login": "^1.0.12",
-    "next": "^12.1.6",
+    "next": "^12.1.7-canary.29",
     "next-auth": "^4.3.4",
     "react": "^18.1.0",
     "react-dom": "^18.1.0"
